Fixed bug with "uncountable" exception while processing floats.
Mathematical conditions produced fatal error with float types, because of incorrect gettype() case.

+++ src/ConditionsRunner.php
@@ -51,7 +51,7 @@ class ConditionsRunner extends RunningSuit
     {
         switch (gettype($value)) {
             case 'integer':
-            case 'float':
+            case 'double':
                 break;
             case 'string':
                 $value = strlen($value);
